Minor changes to particle-system

+++ TheSimulation/TheSimulation/Models/Particle.cs
@@ -3,13 +3,13 @@ using System.Windows.Shapes;
 
 namespace TheSimulation;
 
-public sealed class Particle(Ellipse visual, Vector velocity, double lifetime)
+public sealed class Particle(Ellipse visual, Vector velocity, float lifetime)
 {
     public Ellipse Visual { get; } = visual;
     public Vector Velocity { get; } = velocity;
-    public double Lifetime { get; private set; } = lifetime;
+    public float Lifetime { get; private set; } = lifetime;
 
-    public void DecreaseLifetime(double amount)
+    public void DecreaseLifetime(float amount)
     {
         Lifetime -= amount;
     }
